+++ src/components/AuthModal/AuthModal.tsx
@@ -1,14 +1,12 @@
-import { useEffect, useState } from 'react';
+import { useState } from 'react';
 import Button from 'react-bootstrap/Button';
 import Modal from 'react-bootstrap/Modal';
-import Form from 'react-bootstrap/Form';
-import InputGroup from 'react-bootstrap/InputGroup';
-import { useMutation } from '@apollo/client';
 
-import { useAuth } from '../../context/AuthContext';
+import { HEADER_LABELS, FORM_TYPE } from './AuthModal.utils';
 
-import { LOGIN_MUTATION } from './AuthModal.graphql';
-import { EMAIL_REGEX } from './AuthModal.utils';
+import LoginForm from './components/LoginForm';
+import ForgotPasswordForm from './components/ForgotPasswordForm';
+import RegisterForm from './components/RegisterForm';
 
 import classes from './AuthModal.module.css';
 
@@ -20,116 +18,38 @@ interface Props {
 export const AuthModal = (props: Props) => {
   const { open, handleClose } = props;
 
-  const { setToken } = useAuth();
-
-  useEffect(() => {
-    // Avoid blinking inside inputs because of Boostrap animation
-    setTimeout(() => {
-      setEmail('');
-      setPassword('');
-      setErrors({ emailError: '', passwordError: '' });
-    }, 200);
-  }, [open]);
-
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-
-  const [errors, setErrors] = useState({ emailError: '', passwordError: '' });
-
-  const [login, { loading }] = useMutation(LOGIN_MUTATION);
-
-  const doLogin = () => {
-    const validationErrors = { emailError: '', passwordError: '' };
-    if (!email.match(EMAIL_REGEX)) {
-      validationErrors.emailError = 'Please provide correct email.';
-    }
-
-    if (password === '') {
-      validationErrors.passwordError = 'You need to type your password.';
-    }
-
-    if (validationErrors.emailError || validationErrors.passwordError) {
-      setErrors(validationErrors);
-      return;
-    }
-
-    login({
-      variables: {
-        email,
-        password
-      },
-      onCompleted: (data) => {
-        const {
-          token,
-          user: { name }
-        } = data.login;
-
-        setToken(token, name);
-
-        handleClose();
-      },
-      onError: (error) => {
-        setErrors({
-          emailError: error.message,
-          passwordError: ''
-        });
-      }
-    });
-  };
+  const [formType, setFormType] = useState(FORM_TYPE.LOGIN);
 
   return (
     <Modal show={open} onHide={handleClose} size="lg" centered>
       <Modal.Header closeButton>
-        <Modal.Title>Already a user? Sign in!</Modal.Title>
+        <Modal.Title>{HEADER_LABELS[formType]}</Modal.Title>
       </Modal.Header>
 
       <Modal.Body>
-        <Form className={classes.form}>
-          <Form.Group className="mb-3" controlId="formBasicEmail">
-            <Form.Label>Email address</Form.Label>
-            <InputGroup hasValidation>
-              <Form.Control
-                type="email"
-                placeholder="Enter email"
-                value={email}
-                onChange={(e) => setEmail(e.target.value)}
-                isInvalid={Boolean(errors.emailError)}
-                disabled={loading}
-              />
-              <Form.Control.Feedback type="invalid">{errors.emailError}</Form.Control.Feedback>
-            </InputGroup>
-          </Form.Group>
-
-          <Form.Group className="mb-3" controlId="formBasicPassword">
-            <Form.Label>Password</Form.Label>
-            <InputGroup hasValidation>
-              <Form.Control
-                type="password"
-                placeholder="Password"
-                value={password}
-                onChange={(e) => setPassword(e.target.value)}
-                isInvalid={Boolean(errors.passwordError)}
-                disabled={loading}
-              />
-              <Form.Control.Feedback type="invalid">{errors.passwordError}</Form.Control.Feedback>
-            </InputGroup>
-          </Form.Group>
-        </Form>
-
-        <div className="d-grid gap-2">
-          <Button variant="primary" onClick={doLogin} disabled={loading}>
-            {loading ? 'Logging in...' : 'Login'}
-          </Button>
-          <Button variant="link" disabled={loading}>
-            Forgot your password?
-          </Button>
-        </div>
+        {formType === FORM_TYPE.LOGIN && (
+          <LoginForm openForgotPasswordForm={() => setFormType(FORM_TYPE.FORGOT_PASSWORD)} />
+        )}
+        {formType === FORM_TYPE.FORGOT_PASSWORD && (
+          <ForgotPasswordForm openRegisterForm={() => setFormType(FORM_TYPE.REGISTER)} />
+        )}
+        {formType === FORM_TYPE.REGISTER && <RegisterForm />}
       </Modal.Body>
 
       <Modal.Footer className={classes.footer}>
         <div className="d-grid gap-2">
-          <Button variant="success" type="submit" disabled={loading}>
-            Create new account
+          <Button
+            variant="success"
+            type="submit"
+            onClick={() => {
+              if (formType === FORM_TYPE.LOGIN) {
+                setFormType(FORM_TYPE.REGISTER);
+                return;
+              }
+
+              setFormType(FORM_TYPE.LOGIN);
+            }}>
+            {formType === FORM_TYPE.LOGIN ? 'Create new account' : 'Back to Login'}
           </Button>
         </div>
       </Modal.Footer>
