+++ src/components/AuthModal/components/ForgotPasswordForm.module.css
@@ -1,0 +1,3 @@
+.descriptionLabel {
+  margin-bottom: 2rem;
+}
