+++ src/components/AuthModal/components/RegisterForm.tsx
@@ -1,0 +1,5 @@
+const RegisterForm = () => {
+  return <div>Register form</div>;
+};
+
+export default RegisterForm;
